refactor(hogql): consolidate property type printing logic in printer classes
