+++ posthog/hogql/printer/clickhouse.py
@@ -5,7 +5,7 @@ from uuid import UUID
 
 from django.conf import settings as django_settings
 
-from posthog.schema import PropertyGroupsMode
+from posthog.schema import MaterializationMode, PropertyGroupsMode
 
 from posthog.hogql import ast
 from posthog.hogql.ast import AST
@@ -267,6 +267,37 @@ class ClickHousePrinter(HogQLPrinter):
                     return property_source.has_expr
 
         return None  # nothing to optimize
+
+    def _print_property_type(self, type: ast.PropertyType):
+        if materialized_property_source := self._get_materialized_property_source_for_property_type(type):
+            # Special handling for $ai_trace_id, $ai_session_id, and $ai_is_error to avoid nullIf wrapping for index optimization
+            if (
+                len(type.chain) == 1
+                and type.chain[0] in ("$ai_trace_id", "$ai_session_id", "$ai_is_error")
+                and isinstance(materialized_property_source, PrintableMaterializedColumn)
+            ):
+                materialized_property_sql = str(materialized_property_source)
+            elif (
+                isinstance(materialized_property_source, PrintableMaterializedColumn)
+                and not materialized_property_source.is_nullable
+            ):
+                # TODO: rematerialize all columns to properly support empty strings and "null" string values.
+                if self.context.modifiers.materializationMode == MaterializationMode.LEGACY_NULL_AS_STRING:
+                    materialized_property_sql = f"nullIf({materialized_property_source}, '')"
+                else:  # MaterializationMode AUTO or LEGACY_NULL_AS_NULL
+                    materialized_property_sql = f"nullIf(nullIf({materialized_property_source}, ''), 'null')"
+            else:
+                materialized_property_sql = str(materialized_property_source)
+
+            if len(type.chain) == 1:
+                return materialized_property_sql
+            else:
+                return self._unsafe_json_extract_trim_quotes(
+                    materialized_property_sql,
+                    self._json_property_args(type.chain[1:]),
+                )
+
+        return self._unsafe_json_extract_trim_quotes(self.visit(type.field_type), self._json_property_args(type.chain))
 
     def _get_optimized_materialized_column_equals_operation(self, node: ast.CompareOperation) -> str | None:
         """
